feat(checkbox): Separate the checkbox group from the checkbox into a separate component
